fix: aumentar precisão do cálculo de mL (3 casas)
- Converteu `conversionFactor`, `flowMilliLitres` e `frozenServingMl` para `double` para suportar fator via MQTT com muitas casas decimais.
- Ajustou o cálculo de volume para manter precisão e arredondar o resultado final para 3 casas decimais.
- Atualizou display/HTTP/Web para trabalhar com mL em ponto flutuante.

Trail (detalhamento):
- FlowMeter/FlowMeter.cpp:
  - cálculo de `flowMilliLitres` com `round(...*1000)/1000`.
  - `lastPaintedMl` passou a `double` + comparações com epsilon (0.0005).
  - display de volume usando `%.3f` e buffers ampliados.
- FlowMeter/Mqtt.cpp: parsing de `fatorConversao` como `double`.
- FlowMeter/Http.cpp: `volumeMl` reportado como `double`.
- FlowMeter/WebServer.cpp: exibe `defaultConvFactor` com 8 casas e faz parsing com `atof()`.
- FlowMeter/Config.h + FlowMeter/Config.cpp:
  - `defaultConvFactor` como `double`.
  - migração EEPROM V2→V3 (compatibilidade via `DeviceConfigV2`) e gravação de schema V3.
- FlowMeter/FlowMeter.ino + FlowMeter/FlowMeter.h:
  - atualizações de tipos (`double`) para manter consistência global.

Made-with: Cursor

+++ FlowMeter/FlowMeter.h
@@ -8,16 +8,16 @@
 extern Display display;
 extern volatile byte pulseCount;
 extern volatile long myPulseCount;
-extern unsigned long flowMilliLitres;
+extern double flowMilliLitres;
 extern double totalValue;
 extern double valorMl;
 extern String textHeader;
 extern double saldo;
 extern double quantidade;
-extern float conversionFactor;
+extern double conversionFactor;
 extern String descricao;
 extern bool servingDisplayFrozen;
-extern unsigned long frozenServingMl;
+extern double frozenServingMl;
 extern double frozenServingValue;
 extern bool httpReportPending;
